% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_cov_sim.R
\name{exp_cov_sim}
\alias{exp_cov_sim}
\title{Exponential Covariance Spatial Simulation
Given coordinates, phi, sigma squared, tau squared, a data matrix, and beta values,
a response with an exponential covariance can be simulated}
\usage{
exp_cov_sim(lat, lon, phi, sigma_sq, tau_sq, X, beta)
}
\arguments{
\item{lat}{latitudes of locations of interest}

\item{lon}{longitudes of locations of interest}

\item{phi}{decay parameter}

\item{sigma_sq}{measurement error}

\item{tau_sq}{nugget}

\item{X}{data matrix}

\item{beta}{coefficients}
}
\description{
Exponential Covariance Spatial Simulation
Given coordinates, phi, sigma squared, tau squared, a data matrix, and beta values,
a response with an exponential covariance can be simulated
}
\examples{
exp_cov_sim()
}
\keyword{cats}
